requirements.txt
# Google Drive API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Gemini AI
google-genai>=1.0.0

# Data validation and models
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# PDF processing
pymupdf>=1.23.0
Pillow>=10.0.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Database (sqlite3 is built-in to Python)
